requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
supabase==2.9.0
requests>=2.31.0
httpx==0.27.0
